requirements.txt
PyGithub
prettytable
